PROJECT-5/requirements.txt
streamlit
google-generativeai
pymupdf
speechrecognition
pyaudio
python-dotenv
